move sections of indexer to other files
