+++ go-indexer/indexer/indexer.go
@@ -3,24 +3,13 @@ package indexer
 import (
 	"context"
 	"database/sql"
-	"encoding/json"
 	"fmt"
-	"math/big"
 	"os"
-	"path/filepath"
-	"strings"
 	"sync"
-	"time"
 
-	"github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/accounts/abi"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
-	"github.com/supabase-community/postgrest-go"
 	supa "github.com/supabase-community/supabase-go"
 	"github.com/timewave/vault-indexer/go-indexer/config"
-	"github.com/timewave/vault-indexer/go-indexer/database"
 	event_queue "github.com/timewave/vault-indexer/go-indexer/event-ingestion-queue"
 	"github.com/timewave/vault-indexer/go-indexer/health"
 	"github.com/timewave/vault-indexer/go-indexer/logger"
@@ -256,193 +245,6 @@ func (i *Indexer) Start() error {
 	return nil
 }
 
-func (i *Indexer) setupSubscriptions(wg *sync.WaitGroup) error {
-
-	for _, contract := range i.config.Contracts {
-		parsedABI, err := i.loadAbi(contract)
-		if err != nil {
-			return fmt.Errorf("failed to load ABI for contract %s: %w", contract.Address, err)
-		}
-		contractAddress := common.HexToAddress(contract.Address)
-
-		for _, _eventName := range contract.Events {
-			eventName := parsedABI.Events[_eventName]
-
-			wg.Add(1)
-			go func(event abi.Event) {
-				// Signal that this subscription goroutine has started
-				wg.Done()
-
-				for {
-					// Check if context is cancelled before attempting subscription
-					select {
-					case <-i.ctx.Done():
-						// Don't send error on context cancellation during shutdown
-						return
-					default:
-						// Continue with subscription
-					}
-
-					err := i.subscribeToEvent(contractAddress, event)
-
-					if err != nil {
-						// Use centralized error channel
-						select {
-						case i.errorChan <- fmt.Errorf("failed to subscribe to event %s for contract %s: %w", event.Name, contract.Address, err):
-						default:
-							// Channel is closed or full, ignore
-						}
-						return
-					}
-
-					// Check context again before sleeping
-					select {
-					case <-i.ctx.Done():
-						// Don't send error on context cancellation during shutdown
-						return
-					case <-time.After(2 * time.Second):
-						// Continue to next iteration
-					}
-				}
-			}(eventName)
-		}
-	}
-
-	return nil
-}
-
-func (i *Indexer) loadHistoricalEvents(_currentBlock uint64, wg *sync.WaitGroup) error {
-
-	for _, contractConfig := range i.config.Contracts {
-
-		for _, eventName := range contractConfig.Events {
-			wg.Add(1)
-			parsedABI, err := i.loadAbi(contractConfig)
-			if err != nil {
-				return fmt.Errorf("failed to load ABI for contract %s: %w", contractConfig.Address, err)
-			}
-			event := parsedABI.Events[eventName]
-
-			go func(contractConfig config.ContractConfig, event abi.Event) {
-				defer wg.Done()
-				contractAddress := common.HexToAddress(contractConfig.Address)
-
-				select {
-				case <-i.ctx.Done():
-					return
-				default:
-					// continue
-				}
-
-				lastIndexedBlock, err := i.getLastIndexedBlock(contractAddress, event)
-				if err != nil {
-					// Use centralized error channel
-					select {
-					case i.errorChan <- fmt.Errorf("failed to get last indexed block: %w, event: %s, contract: %s", err, event.Name, contractConfig.Address):
-					default:
-						// Channel is closed or full, ignore
-					}
-					return
-				}
-				if lastIndexedBlock == nil {
-					v := int64(contractConfig.StartBlock)
-					lastIndexedBlock = &v
-				}
-
-				fromBlock := big.NewInt(int64(*lastIndexedBlock))
-				toBlock := big.NewInt(int64(_currentBlock))
-
-				i.logger.Info("Fetching historical events for %s on contract %s, from block %d to block %d", event.Name, contractConfig.Address, *lastIndexedBlock, _currentBlock)
-
-				query := ethereum.FilterQuery{
-					Addresses: []common.Address{contractAddress},
-					Topics:    [][]common.Hash{{event.ID}},
-					FromBlock: fromBlock,
-					ToBlock:   toBlock,
-				}
-
-				logs, err := i.ethClient.FilterLogs(i.ctx, query)
-				if err != nil {
-					// Use centralized error channel
-					select {
-					case i.errorChan <- fmt.Errorf("failed to get historical events: %w, event: %s, contract: %s", err, event.Name, contractConfig.Address):
-					default:
-						// Channel is closed or full, ignore
-					}
-					return
-				}
-
-				for _, vLog := range logs {
-					i.logger.Debug("Received historical event: %v", vLog)
-					eventLog := event_queue.EventLog{
-						BlockNumber:     vLog.BlockNumber,
-						LogIndex:        vLog.Index,
-						Event:           event,
-						Data:            vLog,
-						ContractAddress: contractAddress,
-						BlockHash:       vLog.BlockHash,
-					}
-					i.eventQueue.Insert(eventLog)
-				}
-			}(contractConfig, event)
-		}
-	}
-	return nil
-}
-
-func (i *Indexer) subscribeToEvent(contractAddress common.Address, event abi.Event) error {
-	i.logger.Debug("Subscribing to %s on contract %s", event.Name, contractAddress.String())
-
-	query := ethereum.FilterQuery{
-		Addresses: []common.Address{contractAddress},
-		Topics:    [][]common.Hash{{event.ID}},
-	}
-
-	logs := make(chan types.Log)
-	sub, err := i.ethClient.SubscribeFilterLogs(i.ctx, query, logs)
-	if err != nil {
-		return fmt.Errorf("failed to subscribe to logs: %w", err)
-	}
-
-	for {
-		select {
-		case err := <-sub.Err():
-			i.logger.Error("Subscription error: %v", err)
-			sub.Unsubscribe()
-			return err // Bubble up the error so outer loop can reconnect
-
-		case vLog := <-logs:
-			i.logger.Info("Received log: %v", vLog)
-			eventLog := event_queue.EventLog{
-				BlockNumber:     vLog.BlockNumber,
-				LogIndex:        vLog.Index,
-				Event:           event,
-				Data:            vLog,
-				ContractAddress: contractAddress,
-				BlockHash:       vLog.BlockHash,
-			}
-			i.eventQueue.Insert(eventLog)
-			i.logger.Info("Event inserted into sorted queue: block %d, log index %d", vLog.BlockNumber, vLog.Index)
-		case <-i.ctx.Done():
-			sub.Unsubscribe()
-			return nil
-		}
-	}
-}
-
-func (i *Indexer) loadAbi(contract config.ContractConfig) (abi.ABI, error) {
-	abiPath := filepath.Join("abis", contract.ABIPath)
-	abiBytes, err := os.ReadFile(abiPath)
-	if err != nil {
-		return abi.ABI{}, fmt.Errorf("failed to read ABI file: %w", err)
-	}
-	parsedABI, err := abi.JSON(strings.NewReader(string(abiBytes)))
-	if err != nil {
-		return abi.ABI{}, fmt.Errorf("failed to parse ABI: %w", err)
-	}
-	return parsedABI, nil
-}
-
 func (i *Indexer) Stop() error {
 	i.once.Do(func() {
 		i.logger.Info("Stopping indexer...")
@@ -477,34 +279,6 @@ func (i *Indexer) Stop() error {
 	return nil
 }
 
-func (i *Indexer) getLastIndexedBlock(contractAddress common.Address, event abi.Event) (*int64, error) {
-	data, _, err := i.supabaseClient.From("events").
-		Select("block_number", "", false).
-		Eq("contract_address", contractAddress.String()).
-		Eq("event_name", event.Name).
-		Order("block_number", &postgrest.OrderOpts{Ascending: false}).
-		Limit(1, "").
-		Execute()
-
-	if err != nil {
-		return nil, err
-	}
-
-	var indexedEvents []database.PublicEventsSelect
-
-	if err := json.Unmarshal(data, &indexedEvents); err != nil {
-		return nil, err
-	}
-
-	if len(indexedEvents) == 0 {
-		return nil, nil
-	}
-
-	lastIndexedEvent := indexedEvents[0]
-
-	return &lastIndexedEvent.BlockNumber, nil
-}
-
 func (i *Indexer) startErrorAndReorgMonitoring() {
 	go func() {
 		for {
@@ -523,32 +297,4 @@ func (i *Indexer) startErrorAndReorgMonitoring() {
 			}
 		}
 	}()
-}
-
-func (i *Indexer) handleReorg(reorgEvent ReorgEvent) {
-	i.logger.Info("Handling reorg event: %+v", reorgEvent)
-
-	// 1. Stop all child processes
-	i.logger.Info("Stopping all child processes due to reorg...")
-	i.extractor.Stop()
-	i.transformer.Stop()
-	i.finalityProcessor.Stop()
-
-	// 2. Clear the event queue
-	i.logger.Info("Clearing event queue due to reorg...")
-	i.eventQueue.Clear()
-
-	// 3. Stop the event processor
-	i.logger.Info("Stopping event processor due to reorg...")
-	i.eventProcessor.Stop()
-
-	// 3. Execute cleanup logic
-	i.logger.Info("Executing reorg cleanup logic...")
-	if err := i.reorgHandler.cleanupReorgData(reorgEvent); err != nil {
-		i.logger.Error("Failed to cleanup reorg data: %v", err)
-	}
-
-	// 4. Power down the indexer
-	i.logger.Info("Powering down indexer after reorg handling...")
-	i.Stop()
 }
