+++ go-indexer/indexer/reog-handler.go
@@ -1,0 +1,48 @@
+package indexer
+
+func (i *Indexer) handleReorg(reorgEvent ReorgEvent) {
+	i.logger.Info("Handling reorg event: %+v", reorgEvent)
+
+	// 1. Stop all child processes
+	i.logger.Info("Stopping all child processes due to reorg...")
+	i.extractor.Stop()
+	i.transformer.Stop()
+	i.finalityProcessor.Stop()
+
+	// 2. Clear the event queue
+	i.logger.Info("Clearing event queue due to reorg...")
+	i.eventQueue.Clear()
+
+	// 3. Stop the event processor
+	i.logger.Info("Stopping event processor due to reorg...")
+	i.eventProcessor.Stop()
+
+	// 3. Execute cleanup logic
+	i.logger.Info("Executing reorg cleanup logic...")
+	if err := i.cleanupReorgData(reorgEvent); err != nil {
+		i.logger.Error("Failed to cleanup reorg data: %v", err)
+	}
+
+	// 4. Power down the indexer
+	i.logger.Info("Powering down indexer after reorg handling...")
+	i.Stop()
+}
+
+func (i *Indexer) cleanupReorgData(reorgEvent ReorgEvent) error {
+	i.logger.Info("Cleaning up data for reorg at block %d", reorgEvent.BlockNumber)
+
+	// TODO: Implement specific cleanup logic based on your requirements
+	// This could include:
+	// - Removing events from database that are from the reorged blocks
+	// - Marking positions as invalid
+	// - Clearing cached data
+	// - Updating finality status
+
+	// Example cleanup operations:
+	// 1. Delete events from reorged blocks
+	// 2. Reset position tracking
+	// 3. Clear any cached state
+
+	i.logger.Info("Reorg cleanup completed for block %d", reorgEvent.BlockNumber)
+	return nil
+}
